+++ internal/download/artifact.go
@@ -1,0 +1,21 @@
+package download
+
+import "github.com/saucelabs/saucectl/internal/config"
+
+// ShouldDownload returns true if it should download artifacts, otherwise false
+func ShouldDownloadArtifact(jobID string, passed bool, cfg config.ArtifactDownload) bool {
+	if jobID == "" {
+		return false
+	}
+	if cfg.When == config.WhenAlways {
+		return true
+	}
+	if cfg.When == config.WhenFail && !passed {
+		return true
+	}
+	if cfg.When == config.WhenPass && passed {
+		return true
+	}
+
+	return false
+}
